weatherappai/requirements.txt
requests
PyQt5
python-dotenv
Pillow
